test: move get backfill info tests into new file
